# contracts/nk-genios.openapi-3.1.json
{
  "components": {
    "schemas": {
      "Campus": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "pattern": "^[a-z0-9][a-z0-9-]{1,63}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "Campus",
        "type": "object"
      },
...
      "Envelope_Point_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Point"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[Point]",
        "type": "object"
      },
      "Envelope_list_Campus__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Campus"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[list[Campus]]",
        "type": "object"
      },
...
      "Envelope_list_Point__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[list[Point]]",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorInfo": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo",
        "type": "object"
      },
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Pagination"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "Meta",
        "type": "object"
      },
      "Pagination": {
        "additionalProperties": false,
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "Pagination",
        "type": "object"
      },
...
      "Point": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "campus_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{1,63}$",
            "title": "Campus Id",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/PointCategory"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "summary": {
            "maxLength": 2000,
            "title": "Summary",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "campus_id",
          "name",
          "aliases",
          "category",
          "summary",
          "revision",
          "updated_at"
        ],
        "title": "Point",
        "type": "object"
      },
      "PointCategory": {
        "enum": [
          "public_area",
          "patriotic",
          "academic",
          "residence",
          "dining",
          "commerce",
          "landscape",
          "history"
        ],
        "title": "PointCategory",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "只读查询已公开校园资料。无后台写权限；链接需要用户点击，不会自动控制地图。",
    "title": "Twin NKU 校园公开资料",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/campuses": {
      "get": {
        "description": "列出已启用校区，获取真实 campus_id，不要猜测校区ID。",
        "operationId": "listCampuses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Campus__"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [],
        "summary": "列出已启用校区，获取真实 campus_id，不要猜测校区ID。",
        "tags": [
          "campuses"
        ],
        "x-implementation-status": "implemented",
        "x-module": "M00"
      }
    },
    "/api/v1/campuses/{campus_id}/points": {
      "get": {
        "description": "按校区和名称/别名搜索已公开点位；从返回值获取 point_id。结果过多先澄清。",
        "operationId": "listPoints",
        "parameters": [
          {
            "in": "path",
            "name": "campus_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9][a-z0-9-]{1,63}$",
              "title": "Campus Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PointCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_Point__"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [],
        "summary": "按校区和名称/别名搜索已公开点位",
        "tags": [
          "points"
        ],
        "x-implementation-status": "implemented",
        "x-module": "M00"
      }
    },
...
    "/api/v1/points/{point_id}": {
      "get": {
        "description": "查询已公开点位介绍和版本；资料来源如有记载保留在 summary 中。",
        "operationId": "getPoint",
        "parameters": [
          {
            "in": "path",
            "name": "point_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Point Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Point_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [],
        "summary": "查询已公开点位介绍和版本",
        "tags": [
          "points"
        ],
        "x-implementation-status": "implemented",
        "x-module": "M00"
      }
    },
...
  },
  "servers": [
    {
      "url": "https://2512921.cn"
    }
  ]
}